I think the Flattener is completely done (!), except for one detail with the buffer size interface that I can easily work out with Michal.
So try "java at.dms.kjc.sir.SIRBuilder" to see (what I think is) the
low-ir output of HelloWorld6.

I added an LIRSetTape construct to specify the tapes to the runtime
system.  I'm assuming right now that tapes are stored in the
structures/closures input to functions so that I don't have them
around very much.