You may get this twice (accidently C-c the cvs commit).
While implementing messages in the frontend I found some other issues
that need to be ironed out.  UpDown is the first example of a program with
a stream variable.  Before this update I did not handle stream vars
correctly.  I needed to add a symbol table to the pass that stores
SIRInitStatements, given a JLocalVariable.  I am relying on Kopi to
use a consistent JLocalVariable for a variable (I did not want to
have to deal with scoping).  Now I use the correct stream object
for each use of a stream variable (which can only be a portal reg. or an
add/set* method).  Some other minor issues were dealt with also.

Messaging support has yet to be finalized.