The unroller almost works now.  Right now, the unroller checks that:      *      *  1. the initialization is an assignemnt of a constant to a variable      *      - further, the variable is not declared in the initialization;      *        it is only assigned to there (no "for (int i=...)")      *  2. the condition is a relational less-than test of the var and a const      *  3. the incr is addition or multiplication by a const. (use +=1, not ++)      *  4. the variable is an integer
It does not check that the induction variable is unmodified in the
loop.  This is easy but tedious to add, so for now please don't muck
w/ your loop counters.

The only thing that's preventing the unroller from working end-to-end
is that when stream constructs are replicated, they need to be
registered with their parents (so that the SIRPipeline knows about its
new children.)  This will take another visitor in the SIR, which I'll
write now... (gordo, you don't have to worry about building the
elements list of SIRPipeline's and SIRSplitJoins anymore, since this
pass will rebuild them.)