Modifying the visitor interface for JBlock's.  The internal rep. had changed to a list, but the visitors were still getting an array, which made it look like they could mutate the rep.
So, I'm going out on a limb and breaking with Kopi convention (oh no!)
to NOT send every internal field to a visitor.  From now on, let's
only pass the object to the visitor, and the visitor can get whatever
it wants through the public methods of the object it's visiting.

Otherwise we have cases like this where the rep. changes and then
offers a different view to the visitor, but the visitor doesn't know
about the change since it still thinks it's getting the rep. instead
of just interfacing through the public methods.  It should also save
big-time on having to modify lots of files when you just have dinky
changes!