Now fusion works for everything in the docs/examples directory, fmradio, and perftest.  GSM is the only thing that isn't completely functional--here I get a seg fault running a joiner, and I'll need some help debugging.
Fusion is a serious mess when we're representing the fused filter as a
synthesis of two other filters for all of the scheduling stuff.  We
might want to talk about a cleaner design at some point.

JLocalVariable now has a mutable name, since we needed to mutate
variable names instead of constructing new ones to get the renamer
right.

Kopi2SIR now mutats JVariableDefinitions instead of creating new ones,
since once a new one was created we lost track of all the references
when the definition was renamed.

ReplacingVisitor now includes a visit statement for a variable
declaration statement, which is was missing before.