Adding support for "parameterizable round-robins."  If you say:
	setSplitter(ROUND_ROBIN(N))

the compiler will now treat this as a WEIGHTED_ROUND_ROBIN(N, N, N...)
for the entire width of the split-join.  This is needed for PCADemo,
matrix multiply, and some others, I think.  In the future we can
attack a more general syntax that lets you paramterize the weights,
but this syntax was an easy extension for what we need for now.