SIRBuilder calls EMPTY().
SIRContainer has new <replace> and <contains> methods that subclasses
must implement--replace replaces one stream with another, and contains
returns whether or not a stream is a child of a given stream.

One change I made led to a number of bug fixes -- every time a child
is added to a stream, it now sets that child's parent to be the stream
being added to.  This could be kind of brittle (if you start adding to
your children on the side, forgetting to use your <add> method) but it
fixed things for now.  Might want to revisit a clean model for setting
parents, etc.

Implementing those functions are the changes to Pipeline, SplitJoin,
FeedbackLoop.

TwoStageFilter has init-Work/Pop/Push/Peek mutators.

Flattener has a call to the fission module, but commented out for now.