Re-arranging some of the RAW stuff slightly so that multiple instances of RawFlattener are allowed.  This is useful for me for fission/fussion, but involved changing some static stuff to non-static, which propagated some cahnges into FlatNode.
Hacked around a bug in FlatIRToC that was assuming that local arrays
are allocated at the point of the variable declaration.  When this is
not the case, it had a type error.  Now when it's not the case, it
allocates it on the heap.  We should find a way to allocate on the
stack instead.

RawBackend also now has a call to AdjustGranularity, which does
nothing unless you pass a flag to the java VM which is
-Dapp=<app-name>, e.g. -Dapp=fm.  At least for the next day, I'll be
doing app-specific transformations.