Adding two versions of the BeamFormer that I used for gathering numbers.  BeamFormer-nested1 breaks up the splitjoins into hierarchical splitjoins so that the current version of the partitioner can decrease the width of a splitjoin.  Beamformer-nested2 is a similar hand-partitioning of the beamformer for a certain set of parameters, although it's only of historical interest (the performance was poor, maybe due to routing.)  I'm removing RawBeamFormer because this isn't needed anymore -- it was inserting dummy sources and sinks to avoid a null split or join on RAW.
Generally speaking, we really need to move the programs we're using
for gathering numbers out of /home/bits6 and into source control.  I'm
not sure of the best way to do this... maybe a read-only directory
somewhere with each of the programs for each version of the paper,
plus the command line args that were used and the date that it was
run?  Problem being that the source code isn't always enough to
reconstruct the behavior, as it changes as the compiler improves.