This update changes the IR to interface to the stream hierarchy using the (child, init-args) held in SIRContainer instead of the (target, args) pair held in SIRInitStatements (which were embedded in init functions themselves).  This simplified a lot of ugly code, and will be necessary for the automatic immutable replacement.
Anything with a visitor is changing the signature of the
visitSIRInitStatement call to not include the argument list of the
target.  There are very few cases where anyone should do anything with
an SIRInitStatement anymore; rather, the stream hierarchy and
parameters should be retrieved directly from the stream classes.  (I
don't think this affects any work that people have in flight, though.)