Fixing a few bugs so that FM Radio compiles, runs, & gets the right answer under RAW.
FlatIRToC now generates a function and a call to the initWork function
instead of inlining it in work.  (This allows some necessary variable
declarations in initWork that were being initialized wrong by work's
initializers before, as well as allowing some optimizations like var
decl raising.)

RawBackend now doesn't include the special case fix of the scheduling
bug from the ASPLOS submit, since this no longer occurs with this
FMRadio.

Just changing the name of a variable in renamer (using SIRRenamer as
an instance variable had me searching for class SIRRenamer for 5 mins!)