IR tree constructs that were being recreated from the original parse tree (eg with new JVariableDeclaration(...) were not passing through comments and token references.
I went though and added self.getTokenReference() and self.getComments(), replacing null in the constructors appropriately.

I am hoping to use special comments in the near future for a hack-ish hook into the compiler to let me declare variables as "bit"s, and not have to deal with grammar hacking or screwing with java compatability.