Various changes * Changed the way we send and receive to/from the static network, we now send directly from mem in certain cases
* Changed the translation of println(), I do not know if it is any
faster

* I now recognize certain joiners that are unnecessary (i.e. they do
not buffer data).  This only runs in rate match mode (-x, --ratematch).
The upstream nodes connected to the joiner cannot be joiners, they cannot
execute in the init schedule.  Each incoming weight of the joiner must equal
the push rate times the steady-state execution rate for the filter it is
connected to.

Some other things...