Reduce N_samp from 32 to 6 so the fuser has a hope of fitting this on to 16-tile RAW.  Rewrite source and sink to push an element at a time.  Filters that look like this:
  public void init(int N)
  {
    this.N = N;
    output = new Channel(Something.TYPE, N);
  }
  public void work()
  {
    for (int i = 0; i < N; i++)
      // Do things
  }

are almost always wrong and unnecessarily cripple the scheduler.
(This is also the fundamental bug in the current GSM implementation;
in this case, the scheduler was statically computing the LCM of 2048
and 6 and writing that many unrolled iterations to the C code.)  Write
the work function for a single iteration and use some filter state
instead.