Changed the inline assembly call generated by the Raw backend for receiving and sending data from/to the static network.
This all started when Jasper showed me some joiner code that was
causing problems.  After searching for a while, I noticed that if the
variable we are receive into is register allocated, gcc would produce
incorrect code.  We were producing an instruction to receive to memory,
but the register holding the variable was not being updated after the
receive, only its address on the stack was being updated.

First, I believed that this was a bug in gcc, but it turned out that
the inline assembly was incorrectly specified.  Now it is correct.