FMRadioApp is a copy of LinkedFMTest.java with the following modifications: 1. I renamed it. 2. I added an explicit decimator to the stream graph (I removed it from being mixed up with the lowpass filter. This is less efficient (because we calculate 5 values that we don't actually use), but an explicit decimator would be included in a block diagram of the system, and my optimization is going to combine the low pass filter and decimator again anyways
This is actually an excellent case in point about why linear analysis will be useful. Whomever originally write FMradio was worried about performance so they made the lowpass filter include the decimation information so that the program doesn't calculate useless values. It is cool that I have simplified the LowPassFilter class at the same time as not sacrificing any performance (after my optimization, of course).