Checking in the dynamic programming partitioner.  This thing appears to be totally awesome!  It finds the optimal structured partitioning for all our benchmarks in less than 1 second, and it can be extended to both fuse and fiss (right now it just fuses.)
It isn't turned on by default right now because we still need fusion
of 2-stage filters in pipelines before all that it outputs can be
applied to a stream graph.  If you have to use it you can uncomment a
line in Partitioner.java.

---

SIRIdentity has an init function now (so it can be fused.)

SIRPipeline has some asserts and debugging output.