the partially completed number gatheving code.  This code now will check if we can generate number gathering code for the app.  Right now the app must:
* have one sink
* all prints in the sink
* have at least one print statement in the sink where:
* cannot have a print in control flow (i.e. a print cannot be in an if/then,
while loop, etc.)
* all prints in the sink are in the work function

I will lift the requirement of only one sink shortly.  To support this I
will have to make special versions of some apps.  For example, bitonic has
a print statement in a while loop, we can either run the app with -u 32, or
unroll the prints in the streamit code.

I have to finish the implementation of the actual bC code.