The number gathering implementation.  Right now it only supports apps with a single sink.  I will add support for multiple sinks later tonight.
With the -N option give to the compiler, if possible, the compiler
will generate code to run the application for 5 steady state executions
and it will create a file called results.out with performance numbers
for the run.  It will fail to generate the code if the pass cannot
determine exactly how many prints per steady state execution, (i.e. the
prints are nested in control flow).  It is not too smart right now
but it works for most apps.