Replace streamit_frequency with streamit_fftw here.  My understanding of the way that ld works is that, if you don't reference the entry point in streamit_fftw, the symbols that file references are irrelevant, so this should have no effect unless you're using Andrew's --frequencyreplacement code.
If you are using that code, make sure you *remove* your existing
libstreamit.a (and possibly libstreamit_p.a) and rebuild.  When you
compile StreamIt-generated C code, you need to link with

  -lstreamit -lsrfftw -lsfftw -lm

In the CAG world, the fftw stuff is installed in /usr/uns, and so this
all should magically work.