Adjusting the partitioner record-keeping to keep track of work in fissed nodes.
PartitionRecord is a new bookkeping unit for a partition, keeping
track of work and contents of a partition.  Changing PartitionUtil to
print straight to file when saving statistics; changing
DynamicProgPartitioner to use PartitionRecord's instead of ad-hoc
hashtables and such.

PartitionDot also outputs list of tiles assigned to a fissed filter.