Fixed an implementation issue in the magic network.
This is a bit tough to explain, but let me try.  Every time a tile hit
an instruction with a read from ctsi, the sim called a function in this file
to get which magic fifo it should receive from.  The problem is that every
time this function was called it stepped the receive schedule for that tile.
This should only happen if the receive was successful, i.e. the fifo
it wanted to receive from had data in it.

So I changed btl around to let me know when the receive was successful, so
I can step the receive schedule for the tile.