The implementation of decoupled execution.  Really it is much more...
Only code for tiles assigned to filters get generated.  For these tiles,
all communication is faked as follows:  sends are just assignments to a
global variable (meaning they do nothing), receive turn into reads from
a global variable which always has the value 1 (or 1.0 if a float).

Next, I remove all print statements in the original application and add
a print statement that prints 0 at the start of the work function, and
a prints statment that prints 1 at the end of the work function.