The implementation for the new decoupled execution mode and the new work estimator.
First, the new decoupled execution mode (--decoupled) will generate code
that will run *only* tiles assigned to filters (joiners are not
generated).  These filters will run in a decoupled manner by writing their
output to a variable and reading their network input from a variable.  When
the application is run in this mode, it will run each work function exactly
once and then quit.  It will write the execution tile for each filter tile
in a file called work_est.out.  This mode was very useful for the work
estimator...

The function RawWorkEstimator.estimateWork(SIRFilter) will use the
decoupled execution mode to run the filter's work function on a single
raw tile and return the number of cycles needed or execution.