Removing "badly" named containers and inserting in new containers. Finished and tested implementation of RBTree and added the STL-style wrapper for it. The STL-style interfaces are not nearly completed, but the most common operations are supported, and I'll add more as they are needed (by me or others).
Containers are now named DLList for Doubly Linked List and OSet for
Ordered Set.

As I've stated before, if there are great upheavals about this,
I'm willing to convert the wrappers for these structures to conform
with the Java-style interfaces, but I'd keep the additional functionality
by extending the interfaces. Of course I'd really rather not.