So the snake boot problem turned out to be a known problem with snake boot. And I quote Mike Taylor:
"There's a race condition in the multi-snakeboot code which is
triggered in the following condition:

The last tile in a row of tiles finishes booting before the first tile
on any row above it. So if you still want to use MULTI_SNAKEBOOT
(since it is faster), you can use the vis window and figure out which
row is booting too fast and even it out by throwing in some more code
in that row."

I am not going to figure out which row is booting too fast, so I have
diabled multi_snakeboot for 8x8 configs.  It will take longer to boot,
oh well.