Multidimensional array fix for nokia.  I finally just punted on this problem.
The final problem that I decided not to tackle was assignment of
multidimensional arrays, with the LHS being a non-allocated java array
converted into a ** (for 2 dimensional arrays).  I could search
thru the code and find the assignment of the LHS and then generate my
own allocation based on this future assignment.  I did not
like that solution.  I just seems cheesy.

In the end I would have to calculate my own pointed arithmetic.  I do
not want to do that.

So in the end, the variable localization pass does not run if the
above case is encountered.  Everything else should work fine.