So the bug that has took over my life for the past 4 days is squished!
After using gdb to compare the execution (line by line) of two 7000 line
programs, it turned out the bug was not in remove globals pass.  I traced
it back to a bug in FuseSplit.  When creating the "NewArrayExpression"
for the buffers it created, it would use the InputType of the filter
for both the peek buffer AND the push buffer.  This is incorrect, the
type should be set to the output type for the push buffer.  Whoever
wrote the code, tried to reuse the same code for both the peek buffer
and the push buffer.

Further complicating the problem, was that the inputtype was void (the
fused filter had no input) and the code converted all void types to int
types.  On raw, ints and floats are the same size, so the beamformer
configurations were running fine without the remove globals pass
because the arrays were calloced instead of stack allocated.  So
that lead me off course for a while.

Anyway, my ranting is over.  The bug is fixed and I can now resume my
numbers gather duties.