Now the number gathering pass will work for vocoder.  Vocoder has multiple sinks, with prints in control flow.  So I first fixed the unroller (previous update).  Next I had to rethink my notion of filter synchronization (when the execution of 2 filters are related in the steady state) it is now much more flexible.
Shortly, 2 filters are synchronized in the steady state if there exists
one path from each filter to the least common ancestor of the filters,
where there are no filters that push zero (on the path).

Also, I had to ignore filter that are sinks, but do not print anything.
Before the code kind of assumed all sinks print.  So the code now
searching thru the multiple synchronized sinks and finds one to
generate monitoring code for...