So, in the most recent reg test there were 3 verification failures.  It turns out that they were all caused by the same problem.
The problem is related to the bug I fixed last week.  In the initialization
schedule, if a splitter is scheduled to receive data and hold it in its
incoming buffer there was a problem because I do not map splitters to
tiles.  So there does not exist an incoming buffer for the splitter.  So
I forward the data to the downstream filter.  Last week, I generated tile
code to perform the receive.

But, now the buffer of the filter must be adjusted to account for these
extra items.  That was the bug, the buffers were too small because they
were not accounting for the extra data.  Funny thing:  I looked at my
thesis to correct this problem, and the thesis accounts for it.  But
for some reason, I did not code this up way back.  I must have written
the code before the thesis and never checked to see if it matched.