With this commit rate matching is finally ready for primetime.  It is not enabled by default (use --ratematch) and it only works for layouts with no crossed routes.  The pass with check to see if the layout has no crossed routes and if not, rate matched code / communication will be generated.
Some may ask, what is rate matching?  With rate matching, in the
steady state, each filter will receive enough data to execute all the
work function executions for an entire steady-state.  Then the work
function is executed the number of times the filter appears in the
steady state schedule with the push values buffered.
Then the push buffer is injected into the network.  Of course, all
this is in an infinite loop.  The initialization phase remains
unchanged...

I cannot say that it works for many of the benchmarks we have, but
I looked at FIR and it added about 15% to MFLOPS, not so great, but
enough to write a paper about ;-)