I guess Dave changed something in the front-end because arrays are now generated a bit differently by the front-end.  I used to be able to rely on the fact that arry definition was separate from declaration.  That is not the case anymore.
So I extended the array dimension resolver to handle new array exprs in
variable definitions and I now stack allocate all arrays allocated inside
a variable definition statement.

My array code is very brittle.  It is extremely sensitive to the way Dave
generates the arrays in the IR.  But it has to be this way because of the
incompatibilities between C arrays and Java arrays.