Adjusting the numSamples parameter from 64 to 1024.  This was a long-overdue change, postponed only so we could have backwards compatibility with our benchmarking results.  The other parameters we had were all exactly scaled down from the specification by a factor of 4, except for numSamples, which used to be a factor of 16.  Now it is a factor of 4 like the rest, so this actually makes sense.
The only bad side effect is that the Coarse versions get too big to be
usable by linear analysis.  Not sure what the best thing to do about
this is, besides making Andrew's stuff more efficient.  The plain
version runs in reasonable time as before.

Also adjusting the output files, and removing the Coarse one since it
is identical to the plain one.