If --outputs broke --simulatework.  Output limiting would be enabled for profiling and it would never reach the appropiate number of outputs.
It is fixed now by setting --outputs to -1 in MakefileGenerator.java.

It should be fine now, but I do not think that we should use --outputs for
the regtest just yet (I did even know it was used in the regtest yet).
I want to add deadlock detection to it.  That should not take too long.