Here is something that I have had implemented for a while but I was a bit scared to check in:
Now the number gathering infrastructure *should* die when the
chip is deadlocked.  If no new instruction has issue on any
tile mapped to a filter or joiner for X cycles (X is now 2000),
then die and report it in the results.out file.

I have been using this a while and there are no known issues with it.
I checked it in because it seems like Jeremy can really use this.  He is
really pushing the old backend by compiling to some large (> 4x4 layouts).
And some of the runs are taking a really long time.