In the library backend, don't use 'new' to create objects; instead, create fields for each init parameter, and create a public static __construct() method that saves its parameters away, and override callInit() to call init() with the saved parameters.  (Should check if this duplicates work done in MoveStreamParameters so that parameters get saved multiple times.)
In the compiler backend, do use 'new' to create objects, as before,
but generate an empty constructor with the right parameter list.  Kopi
cares that the constructor exists but the compiler ignores its
contents.

This commit should get rid of our dependency on having init function
signatures in the library.