Lots o' changes...
* various bug fixes.
* corrected notion of a redundant buffer.
* noticed that incoming buffers of an input trace cannot share
a dram.  this might kill us.  It means that, if an output node is
connected to an input node, the total number of streams split/joined
by each must not exceed the number of ports (16 for a 4x4).  This
may be a problem. we may have to use the old partitioning stuff to
enforce this.