new parameters for run_cluster: -runccp -ccp <hostname or ip address> -init <iteration#>
added code that loads the saved state of a thread
if requested.

moved a call to init() function for filters from
__CLUSTERMAIN to run method.

changed the __CLUSTERMAIN function for filters to
only execute the initialization executions if
starting the execution from steady state 0.

chages in code due to change in node_server library class.