Adding frontend support for static array initializers, such as:
int[2][2] A = { {1, 2}, {3, 4} };

It currently works through to the Java library, though the compiler
still needs some work.  You should be able to declare such arrays as
fields or locals, of any type and number of dimensions.

The only restriction is that you currently have to specify all of the
base elements of the array.  You can't have symbolic sub-arrays in the
initialization, like this:

int[2] B = {1, 2};
int[2][2] C = {B, B};

The main challenge with supporting this is semantic checking (and
probably constant prop in the compiler.)  Please let me know if that
is an important feature (e.g., for the IBM study).