Added:
ConvertArrayInitializers:  A pass that *partially* converts static array
initializers to assignment statements.  It generates the assignments
and relies on another class (flatirtors) to place them in the code.

NewArrayExprs: A class that searches for a new array expression to
associate with each array variable.

VarsAssigned was moved to common.

FlatIRToRS now does all of the array translation necessary for rstream.