Added:
RemoveDeadDoLoops:  Partitioning introduces a lot of for loops that
never execute or execute exactly once (static loop bounds).  So just
remove a for loop that never executes (as long as its init, cond, incr don't
have side-effects).  And remove the header for a loop that executes once and
doesn't use the induction var in the body.

StructureIncludeFile: Generate c struct definitions for all structures used
in the program.

FlatIR calls removedeaddoloops