Jasp told me that all the raw tests deadlock in the reg test. So...
Turn off hardware icaching because the print service does not work with
hw icaching and this is the reason why all the raw tests fail in the reg
test.

This is a short term solution.  In the future we should use
raw_test_pass_reg() to print.  But this prints in a crazy format that would
need to be parsed by the reg test.