Due to the manual layout of the graphics application, XY routing is not adequate.
Adding a more intelligent router that will route around occupied tiles
and find the route with the least amount of occupied tiles.  It always makes
progress toward the destination.  It will only be called if there are
enough free tiles in the layout to warrant it.  Right now I have this set
to 15% free tiles.

It remains to be seen if this will greatly slow down the compiler.