spacedynamic performance results.  run the application with -N <int> where <int> is the number of prints per "steady state" of the application.
Run the application using 'make -f Makefile run' and then inspect the
generated file 'results.out' that is created in the application directory.
The application is run for as long as it takes to generate 10 * <int> outputs.
So it runs for 10 steady states.  Performance for each steady state is
recorded.