Adding the --malloczeros option.  This option informs the various backends that the simulator initializes memory with zeros, so that it is fine to use malloc instead of calloc.  This greatly speeds up initialization time for the graphics apps.
For now this only works with my version of the simulator.  I'm going to
speak with Mike Talyor about incorporating into the next btl release.
Until then, I print a warning to the user if this option is enabled.  If
you want to use this option now, point your $TOP_DIR to
/home/bits6/mgordon/starsearch (change in streams/include/dot-bashrc, etc).