tweaking the cost function and adding a memory cost component to the cost function that will try to place filters with a large memory requirement (one that exceeds the dcache) at near the border of the chip.
Right now the memory cost component is a bit hacked up, it just is a proof
of concept.  But the algorithm beautifully lays out the one graphics app
I have.

When Janis releases his functions for estimating the data footprint of a
filter, I will remove the hacks that target this specific graphics app.