The file device implementation for the space dynamic backend.
It is much more flexible, clean, and robust than before.  You can
attach a file reader or a file writer to any border tile of the chip.
Right now the user assigns file devices in the stream graph to ports
on the chip.  Split file reader and joined file writers are supported.

Note the file reader reads files in binary format (not ascii) and the
filewriter writes in binary.  For each value the file writer writes,
I print the tile coordinates, the cycle, and the hex value to the
screen.

Can't think of anything else.