Fixing 2 bugs in Propagator.  Both appeared when the write variable was false, meaning the propagator should not alter the code but it incorrectly did.
* if an argument to a function resolved to a constant when analyzed,
the propagator replaced it even if write was false

* same thing for the dimensions in a New Array Expression

In the jpeg encoder, the first bug appeared when we examined the body
of a loop to discover the variables that are altered and should not be
resolved to constants.  In its search, it would erroneously replace
arguments of functions that were deemed to be constants at the start
of the loop, but were these variables were in fact altered in the loop.