Allow three new attributes on impl in a benchmarks.xml file:
compile_time="3600"  test compile time in seconds, in this example
allowing an hour rather than the 20 minute default.

run_time = "120" test run time in seconds, in this example allowing 2 minutes
rather than the 20 minute default.  In case of compile-and-go backends
(i.e. library) compile_time changes the time limit and run_time has no effect.

iteratons="1" changes number of strady-state iterations in a test run, in this
example from the default of 100 to 1.