Rearranging the space dynamic backend a bit.  Before the backend only support filters that did not peek from dynamic rate input channels.  So,
BufferedStaticCommunication is a pass that will generate SIR code for a filter
that implements circular or linear buffer input communication over the static
network.  It is working.

BufferedDynamicCommunicatin is a stub right now that will generate SIR code
to implement circular buffer input communication over the dynamic network
(dynamic rate) for a filter.

Also, starting to add support of dynamic rate FileReaders/Writers meaning the
file* is connected to a dynamic rate.