Reworking the logical mapping of drams to tiles.  There are now two mappings in the compiler, the hardware mapping of tiles to drams, and the logical mapping from tiles to drams that they "own", meaning that all commands to the dram owned must issue from the owning tile.
So we generate code for using the logical mapping, with an initial
pass to communicate addresses all over the chip from hw-mapped tile
to logical mapped tile.

The logical mapping used to be that the dram was owned by its neighboring
border tile, but I needing something more complex to support loads over
the gdn.