If we need to store data using the gdn from tile that is not the owner of the target dram, then we have to have synchronization between the tile owner of the dram and the storing tile.
if the owner tile is used by the trace that is storing, we get free
synchronization, otherwise we have to explicitedly synchronize the two
tiles:  send a word over the static network from the owner tile to the
storing tile, that will let the storing tile know the target dram is ready
to receive data for the store command.  Yeah.