on the right track, within 15% of hand scheduled for beamformer.
disregard the any calculation of communication cost, it is meaningless.

bias the work estimation of non-bottleneck tiles so that we account
for error in the estimator.  so make sure that the schedule accounts
for the poor estimate by over-estimating the work of any filter not
mapped to a bottleneck tile...