Adding support for destorying of arrays for the space time backend.
I first attempted to factor out all of the common code between the filter
case and the code store (spacetime) case, but it was time consuming and
not worth it.

In the end I just simply copied code from the SIRFilter case.

I was thinking that we might need a superclass to encapsulate classes that
have methods and fields (code), then the optimizing passes could be more
general.