not so quick, but dirty, fix to following problem:
static {
  int N;
  int[N] myArray;
  init { N = 10;}
}
should (and now does) create an array of constant length 10 at use sites
rather than an array with length given by a variable.

The problem was that the fields for a static section are marked public static
but the reference to N in int[N] myArray was just  a NameExpression when
it should be a FieldAccessExpression with a prefix giving the static section
name.

Once a prepass fixes references to the from the static fields to the static fields
ConstanProp propagates the value correctly.

The fix is "dirty" in several ways:
(1) I had to turn off errors on references to "this" in static contexts,
but still left error message and stack dump.
(2) The proper place to fix is in front end for all static fields, but I fixed
in StaticsProp (which I believe is the only place where the frontend
creates static fields).
(3) Extending the fix, in good Java style to referring to static fields in
methods as "TheGlobals.foo" rather than "foo" causes ConstantProp
to not propagate, but did not track down ConstantProp's fragility around
the issue of explicitly-prefixed fields.