Fix streamit.py to not ignore some error codes: expect some more failures reported next regtest.
Change run-reg-tests.py to enable getting number of cpus to use based on /proc/cpuinfo.

Change build-qmtest to accept a prefix to apply to source files being copied into test directory.  This seemed simplest hack to allow me to run multiple tests from qmtest simultaneously: just create a subdirectory to put apps.qms and QMTest into and cd to the subdirectory to run build-qmtest giving it "../" as a prefix.

Change streamitqm to remove a feature probably not used since D. Maze used it for testing his python code.

P.S. If anyone cares, qmtest's definition of a "test database" is [or is at least as general as] a directory with a subdirectory with a '.qms' extension and a subdirectory named 'QMTest'.  The definition was not obvious and this took a number of tries to find out.