dynamic rates for cluster_standalone, and varous other.
the following incomplete code should be fixed up later:
(1) JEmittedText statement allows arbitrary text to be output from a back end.
It was put in to allow me to put in a fancy initialization of a dynTape in
an init function.  Then useq Q&D method of initializing in fusion.h instead.

(2) Added "attribetes" -- mape to SIROperator as a Q&D way to get around
some missing features of StreamGraph.

The rest is mostly dynamic rates.  (Gee: the prototype looked reasonably small.
How come making it work took a week of coding and debugging?)