Found out the bug that caused the loss of precision in the previous version, the inverse quantization scaling was actually incorrect, so it wasn't actually the loss of precision from float->ints that was causing the problem.
This is now a stable version for filters that can do the following:

1. core transform
2. scale and quantize
3. rescale and quantize
4. inverse transform

There is also an implementation for a supporting the Hadamard transform done on the DC coefficients of each 4x4 block (luma or chroma) within a macroblock. Result is a 4x4 Hadamard Transform on the luma blocks within a macroblock, and also a 2x2 Hadamard Transform on the chroma blocks within a macroblock.

This version only implements a working version of the algorithm. Note that there is an algorithm that only uses right and left shifts.