This is a stable version of filters needed to figure out what the best prediction mode is for 16x16 macroblocks.
It currently allows you to put in a block, and prints out residual values at different stages, and output the best mode, mode that minimizes the sum of absolute errors.


The prediction done in units of 16x16 Macroblocks, and assumes and input with the original pixels + the top, left edge pixels.

All 4 modes (vertical, horizontal, dc, plane) modes are supported.

These filters are to be integrated with larger structures in IntraPrediction.str