(1) Fixed problem in ThreeAddressCode: test for whether to convert "for" statement would come back positive if there was a JExpressionListStatement in the "for" header.  Changed so that JExpressionListStatement with list length 1 does not get converted by default, which fixes the problem for the "for" statements that people care about: the unroller works as well as before.
(2) in testing (1) noticed that there were bugs in SimplifyPopPeekPush.  Looks like intertwining decision as to what to convert to three-address code with actual conversion was too complex and led to bugs.  Made selecting what to convert a separate pass: extensive reworking of ThreeAddressCode, and new visitor ThreeAddressExpressionCheck.  Tests OK on MP3.str, except oddity that changing some expressions to three-address-code causes the a few output values to be off by 1. (Is it the case that floating point computations are carried out at higher precision inside an expression than when keeping intermediate results?  That is the only explanation I can think of.)